rag-mentor-python-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
tiktoken
